+++ bin/session.ml
@@ -28,6 +28,7 @@ module State = struct
     sort_by_no_score : Sort_by.t;
     screen_split : Command.screen_split;
     show_bottom_right_pane : bool;
+    show_key_binding_info_pane : bool;
     focus_list : string list;
     path_highlights : Int_set.t String_map.t;
   }
@@ -67,6 +68,7 @@ module State = struct
         |> (fun (typ, order) -> ((typ :> Sort_by.typ), order));
       screen_split = `Even;
       show_bottom_right_pane = true;
+      show_key_binding_info_pane = true;
       focus_list = [];
       path_highlights = String_map.empty;
     }
@@ -89,6 +91,7 @@ module State = struct
   let show_pane (t : t) (pane : Command.pane) =
     match pane with
     | `Bottom_right -> t.show_bottom_right_pane
+    | `Key_binding_info -> t.show_key_binding_info_pane
 
   let refresh_search_results pool stop_signal (t : t) : t option =
     let cancellation_notifier = Atomic.make false in
@@ -549,6 +552,7 @@ module State = struct
         sort_by_no_score = t.sort_by_no_score;
         screen_split = t.screen_split;
         show_bottom_right_pane = t.show_bottom_right_pane;
+        show_key_binding_info_pane = t.show_key_binding_info_pane;
         focus_list = t.focus_list;
         path_highlights = t.path_highlights;
       }
@@ -567,6 +571,7 @@ module State = struct
           sort_by_no_score = t.sort_by_no_score;
           screen_split = t.screen_split;
           show_bottom_right_pane = t.show_bottom_right_pane;
+          show_key_binding_info_pane = t.show_key_binding_info_pane;
           focus_list = t.focus_list;
           path_highlights = t.path_highlights;
         }
@@ -786,6 +791,7 @@ let run_command pool (command : Command.t) (st : State.t) : (Command.t * State.t
       let st =
         match pane with
         | `Bottom_right -> { st with show_bottom_right_pane = false }
+        | `Key_binding_info -> { st with show_key_binding_info_pane = false }
       in
       Some (command, st)
     )
@@ -793,6 +799,7 @@ let run_command pool (command : Command.t) (st : State.t) : (Command.t * State.t
       let st =
         match pane with
         | `Bottom_right -> { st with show_bottom_right_pane = true }
+        | `Key_binding_info -> { st with show_key_binding_info_pane = true }
       in
       Some (command, st)
     )